# artifacts/api-server/src/routes/products.ts
import { Router } from "express";
import { eq } from "drizzle-orm";
import { db, productsTable, categoriesTable } from "@workspace/db";
import { requireAuth } from "../middlewares/requireAuth.js";
import { logAudit } from "../lib/audit.js";

const router = Router();

router.get("/products", requireAuth, async (req, res): Promise<void> => {
  const rows = await db.select({
    id: productsTable.id,
    name: productsTable.name,
    sku: productsTable.sku,
    categoryId: productsTable.categoryId,
    categoryName: categoriesTable.name,
    unitPrice: productsTable.unitPrice,
    costPrice: productsTable.costPrice,
    stock: productsTable.stock,
    locationId: productsTable.locationId,
    unit: productsTable.unit,
    isActive: productsTable.isActive,
    createdAt: productsTable.createdAt,
  }).from(productsTable)
    .leftJoin(categoriesTable, eq(productsTable.categoryId, categoriesTable.id))
    .orderBy(productsTable.name);
  res.json(rows.map(r => ({ ...r, createdAt: r.createdAt.toISOString() })));
});

router.post("/products", requireAuth, async (req, res): Promise<void> => {
  const { name, sku, categoryId, unitPrice, costPrice, stock, locationId, unit } = req.body as {
    name?: string; sku?: string | null; categoryId?: number | null;
    unitPrice?: string; costPrice?: string; stock?: number; locationId?: number | null; unit?: string;
  };
  if (!name || !unitPrice || !costPrice || !unit) { res.status(400).json({ error: "name, unitPrice, costPrice, unit required" }); return; }
  const [row] = await db.insert(productsTable).values({
    name, sku: sku ?? null, categoryId: categoryId ?? null, unitPrice, costPrice,
    stock: stock ?? 0, locationId: locationId ?? null, unit,
  }).returning();
  await logAudit(req.userId, "create", "product", row!.id, `Created product ${name}`);
  res.status(201).json({ ...row!, categoryName: null, createdAt: row!.createdAt.toISOString() });
});

router.patch("/products/:id", requireAuth, async (req, res): Promise<void> => {
  const id = parseInt(Array.isArray(req.params.id) ? req.params.id[0]! : req.params.id!, 10);
  const { name, sku, categoryId, unitPrice, costPrice, stock, unit, isActive } = req.body as {
    name?: string; sku?: string | null; categoryId?: number | null;
    unitPrice?: string; costPrice?: string; stock?: number; unit?: string; isActive?: boolean;
  };
  const updates: Record<string, unknown> = {};
  if (name != null) updates.name = name;
  if (sku !== undefined) updates.sku = sku;
  if (categoryId !== undefined) updates.categoryId = categoryId;
  if (unitPrice != null) updates.unitPrice = unitPrice;
  if (costPrice != null) updates.costPrice = costPrice;
  if (stock != null) updates.stock = stock;
  if (unit != null) updates.unit = unit;
  if (isActive != null) updates.isActive = isActive;
  const [row] = await db.update(productsTable).set(updates).where(eq(productsTable.id, id)).returning();
  if (!row) { res.status(404).json({ error: "Product not found" }); return; }
  await logAudit(req.userId, "update", "product", id);
  res.json({ ...row, categoryName: null, createdAt: row.createdAt.toISOString() });
});

router.delete("/products/:id", requireAuth, async (req, res): Promise<void> => {
  const id = parseInt(Array.isArray(req.params.id) ? req.params.id[0]! : req.params.id!, 10);
  const [row] = await db.delete(productsTable).where(eq(productsTable.id, id)).returning();
  if (!row) { res.status(404).json({ error: "Product not found" }); return; }
  res.sendStatus(204);
});

export default router;