# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ERP Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all locations
      responses:
        "200":
          description: Locations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Accounts
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: locationId
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Customers
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List all purchases
      parameters:
        - in: query
          name: locationId
          schema:
            type: integer
        - in: query
          name: supplierId
          schema:
            type: integer
      responses:
        "200":
          description: Purchases list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get a purchase by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"

  # Expenses
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Credits
...
components:
  schemas:
...
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, isActive, createdAt]
    CreateLocationRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
      required: [name]
    UpdateLocationRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean

    # Account
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        unitPrice:
          type: string
        costPrice:
          type: string
        stock:
          type: integer
        locationId:
          type: ["integer", "null"]
        unit:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, unitPrice, costPrice, stock, unit, isActive, createdAt]
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        unitPrice:
          type: string
        costPrice:
          type: string
        stock:
          type: integer
        locationId:
          type: ["integer", "null"]
        unit:
          type: string
      required: [name, unitPrice, costPrice, stock, unit]
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        unitPrice:
          type: string
        costPrice:
          type: string
        stock:
          type: integer
        unit:
          type: string
        isActive:
          type: boolean

    # Customer
...
    PurchaseItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: integer
        unitCost:
          type: string
        total:
          type: string
      required: [productId, productName, qty, unitCost, total]

    # Purchase
    Purchase:
      type: object
      properties:
        id:
          type: integer
        invoiceNo:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
        accountId:
          type: ["integer", "null"]
        userId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        subtotal:
          type: string
        discount:
          type: string
        total:
          type: string
        amountPaid:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, invoiceNo, userId, items, subtotal, discount, total, amountPaid, status, createdAt]
    CreatePurchaseRequest:
      type: object
      properties:
        supplierId:
          type: ["integer", "null"]
        locationId:
          type: ["integer", "null"]
        accountId:
          type: ["integer", "null"]
        userId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              qty:
                type: integer
              unitCost:
                type: string
            required: [productId, qty, unitCost]
        discount:
          type: string
        amountPaid:
          type: string
        notes:
          type: ["string", "null"]
      required: [userId, items, discount, amountPaid]

    # Expense
    Expense:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        accountId:
          type: ["integer", "null"]
        accountName:
          type: ["string", "null"]
        userId:
          type: integer
        notes:
          type: ["string", "null"]
        date:
          type: string
        createdAt:
          type: string
      required: [id, title, amount, userId, date, createdAt]
    CreateExpenseRequest:
      type: object
      properties:
        title:
          type: string
        amount:
          type: string
        categoryId:
          type: ["integer", "null"]
        accountId:
          type: ["integer", "null"]
        userId:
          type: integer
        notes:
          type: ["string", "null"]
        date:
          type: string
      required: [title, amount, userId, date]
    UpdateExpenseRequest:
      type: object
      properties:
        title:
          type: string
        amount:
          type: string
        notes:
          type: ["string", "null"]

    # Credit
...